requirements.txt
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
asgiref==3.5.0
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.2.2
fastapi==0.112.2
Flask==3.0.3
h11==0.14.0
httplib2==0.20.2
idna==3.8
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==3.0.2
more-itertools==8.10.0
netifaces==0.11.0
numpy==2.1.2
oauthlib==3.2.0
paramiko==3.5.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
psycopg2==2.9.10
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.6.1
PyGObject==3.42.1
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
python-dotenv==1.0.1
pytz==2024.2
PyYAML==5.4.1
requests==2.32.3
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
systemd-python==234
termcolor==2.5.0
tqdm==4.67.0
typing_extensions==4.12.2
tzlocal==5.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.3
uvicorn==0.30.6
wadllib==1.3.6
Werkzeug==3.0.4
wsproto==1.0.0
zipp==1.0.0
